+++ packages/plait/src/board/board.component.ts
@@ -22,34 +22,34 @@ import { RoughSVG } from 'roughjs/bin/svg';
 import { fromEvent, Subject } from 'rxjs';
 import { filter, takeUntil } from 'rxjs/operators';
 import { SCROLL_BAR_WIDTH } from '../constants';
-import { BaseCursorStatus } from '../interfaces';
-import { PlaitBoard, PlaitBoardChangeEvent, PlaitBoardOptions } from '../interfaces/board';
+import { PlaitBoard, PlaitBoardChangeEvent, PlaitBoardOptions, PlaitBoardViewport } from '../interfaces/board';
 import { PlaitElement } from '../interfaces/element';
 import { PlaitPlugin } from '../interfaces/plugin';
 import { Viewport } from '../interfaces/viewport';
 import { createBoard } from '../plugins/create-board';
 import { withBoard } from '../plugins/with-board';
 import { withHistory } from '../plugins/with-history';
-import { withMove } from '../plugins/with-move';
+import { withHandPointer } from '../plugins/with-hand';
 import { withSelection } from '../plugins/with-selection';
 import { Transforms } from '../transforms';
 import {
-    calculateBBox,
-    calculateZoom,
-    convertViewport,
-    getViewportCanvasBox,
-    getViewportClientBox,
-    invert,
-    invertViewport,
+    getRootGroupBBox,
+    clampZoomLevel,
+    convertToViewportCoordinates,
+    getViewportContainerBox,
+    invertViewportCoordinates,
     transformMat3,
-    updateCursorStatus
+    getBoardClientBox
 } from '../utils';
 import { BOARD_TO_ON_CHANGE, HOST_TO_ROUGH_SVG, IS_TEXT_EDITABLE, PLAIT_BOARD_TO_COMPONENT } from '../utils/weak-maps';
+import { BoardComponentInterface } from './board.component.interface';
+import { RectangleClient } from '../interfaces/rectangle-client';
+import { PlaitPointerType } from '../interfaces/pointer';
 
 @Component({
     selector: 'plait-board',
     template: `
-        <div class="viewport-container" #container>
+        <div class="viewport-container" #viewportContainer>
             <svg #svg width="100%" height="100%" style="position: relative;"></svg>
             <plait-element
                 *ngFor="let item of board.children; let index = index; trackBy: trackBy"
@@ -62,9 +62,7 @@ import { BOARD_TO_ON_CHANGE, HOST_TO_ROUGH_SVG, IS_TEXT_EDITABLE, PLAIT_BOARD_TO
         </div>
         <plait-toolbar
             *ngIf="isFocused && !toolbarTemplateRef"
-            [cursorStatus]="board.cursor"
-            [viewZoom]="board.viewport!.zoom"
-            (moveHandle)="changeMoveMode($event)"
+            [board]="board"
             (adaptHandle)="adaptHandle()"
             (zoomInHandle)="zoomInHandle()"
             (zoomOutHandle)="zoomOutHandle()"
@@ -74,7 +72,7 @@ import { BOARD_TO_ON_CHANGE, HOST_TO_ROUGH_SVG, IS_TEXT_EDITABLE, PLAIT_BOARD_TO
     `,
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, OnDestroy {
+export class PlaitBoardComponent implements BoardComponentInterface, OnInit, OnChanges, AfterViewInit, OnDestroy {
     hasInitialized = false;
 
     board!: PlaitBoard;
@@ -83,33 +81,14 @@ export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, On
 
     destroy$: Subject<any> = new Subject();
 
-    public autoFitPadding = 8;
+    viewportState: PlaitBoardViewport = {
+        zoom: 1,
+        autoFitPadding: 8
+    };
 
-    public isMoving: boolean = false;
-
-    private zoom = 1;
-
-    private viewportCanvasBox = {};
-
-    private canvasWidth!: number;
-
-    private canvasHeight!: number;
-
-    private viewBox!: number[];
-
-    private focusPoint: number[] = [0, 0];
-
-    private matrix!: number[];
+    public isMoving = false;
 
     private resizeObserver!: ResizeObserver;
-
-    width = 0;
-
-    height = 0;
-
-    scrollLeft!: number;
-
-    scrollTop!: number;
 
     @Input() plaitValue: PlaitElement[] = [];
 
@@ -123,10 +102,6 @@ export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, On
 
     @Output() plaitBoardInitialized: EventEmitter<PlaitBoard> = new EventEmitter();
 
-    public get isMoveMode(): boolean {
-        return this.board.cursor === BaseCursorStatus.move;
-    }
-
     get host(): SVGElement {
         return this.svg.nativeElement;
     }
@@ -137,7 +112,7 @@ export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, On
 
     @HostBinding('class')
     get hostClass() {
-        return `plait-board-container ${this.board.cursor}`;
+        return `plait-board-container ${this.board.pointer}`;
     }
 
     @HostBinding('class.readonly')
@@ -147,7 +122,7 @@ export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, On
 
     @HostBinding('class.moving')
     get moving() {
-        return this.board.cursor === BaseCursorStatus.move && this.isMoving;
+        return this.board.pointer === PlaitPointerType.hand && this.isMoving;
     }
 
     @HostBinding('class.focused')
@@ -161,13 +136,15 @@ export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, On
     @ContentChild('plaitToolbar')
     public toolbarTemplateRef!: TemplateRef<any>;
 
-    @ViewChild('container', { read: ElementRef, static: true })
-    contentContainer!: ElementRef;
+    @ViewChild('viewportContainer', { read: ElementRef, static: true })
+    viewportContainer!: ElementRef;
 
     constructor(public cdr: ChangeDetectorRef, private renderer2: Renderer2, private elementRef: ElementRef) {}
 
     ngOnInit(): void {
-        const roughSVG = rough.svg(this.host as SVGSVGElement, { options: { roughness: 0, strokeWidth: 1 } });
+        const roughSVG = rough.svg(this.host as SVGSVGElement, {
+            options: { roughness: 0, strokeWidth: 1 }
+        });
         HOST_TO_ROUGH_SVG.set(this.host, roughSVG);
         this.initializePlugins();
         this.initializeEvents();
@@ -180,9 +157,6 @@ export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, On
                 viewport: this.board.viewport,
                 selection: this.board.selection
             };
-            if (this.board.operations.length > 0) {
-                this.updateViewport();
-            }
             this.plaitChange.emit(changeEvent);
         });
         this.hasInitialized = true;
@@ -203,19 +177,23 @@ export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, On
 
     ngAfterViewInit(): void {
         this.plaitBoardInitialized.emit(this.board);
-        this.initContainerSize();
-        this.setViewport(1);
-        this.plaitViewport ? this.initCanvas(this.plaitViewport) : this.focusGraphicCenter();
+        this.initViewportContainer();
+        this.calcViewBox(this.viewportState.zoom);
+        this.initViewport();
     }
 
     private initializePlugins() {
-        let board = withMove(withHistory(withSelection(withBoard(createBoard(this.host, this.plaitValue, this.plaitOptions)))));
+        let board = withHandPointer(withHistory(withSelection(withBoard(createBoard(this.host, this.plaitValue, this.plaitOptions)))));
         this.plaitPlugins.forEach(plugin => {
             board = plugin(board);
         });
         this.board = board;
+
         if (this.plaitViewport) {
             this.board.viewport = this.plaitViewport;
+            this.updateViewportState({
+                zoom: this.plaitViewport?.zoom ?? 1
+            });
         }
     }
 
@@ -230,6 +208,12 @@ export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, On
             .pipe(takeUntil(this.destroy$))
             .subscribe((event: MouseEvent) => {
                 this.board.mousemove(event);
+            });
+
+        fromEvent<MouseEvent>(this.host, 'mouseup')
+            .pipe(takeUntil(this.destroy$))
+            .subscribe((event: MouseEvent) => {
+                this.board.mouseup(event);
             });
 
         fromEvent<MouseEvent>(document, 'mouseup')
@@ -302,29 +286,7 @@ export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, On
                 this.board?.deleteFragment(event.clipboardData);
             });
 
-        fromEvent<MouseEvent>(this.contentContainer.nativeElement, 'mousemove')
-            .pipe(
-                takeUntil(this.destroy$),
-                filter(() => {
-                    return !!this.isFocused;
-                })
-            )
-            .subscribe((e: MouseEvent) => {
-                this.focusPoint = [e.clientX, e.clientY];
-            });
-
-        fromEvent<MouseEvent>(this.contentContainer.nativeElement, 'mouseleave')
-            .pipe(
-                takeUntil(this.destroy$),
-                filter(() => {
-                    return !!this.isFocused;
-                })
-            )
-            .subscribe((e: MouseEvent) => {
-                this.resetFocusPoint();
-            });
-
-        fromEvent<MouseEvent>(this.contentContainer.nativeElement, 'scroll')
+        fromEvent<MouseEvent>(this.viewportContainer.nativeElement, 'scroll')
             .pipe(
                 takeUntil(this.destroy$),
                 filter(() => {
@@ -332,11 +294,14 @@ export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, On
                 })
             )
             .subscribe((event: Event) => {
-                const scrollLeft = (event.target as HTMLElement).scrollLeft;
-                const scrollTop = (event.target as HTMLElement).scrollTop;
-
-                (Math.abs(this.scrollLeft - scrollLeft) <= 1 && Math.abs(this.scrollTop - scrollTop) <= 1) ||
-                    this.setScroll(scrollLeft, scrollTop);
+                const { scrollLeft, scrollTop } = event.target as HTMLElement;
+                const left = this.viewportState.scrollLeft!;
+                const top = this.viewportState.scrollTop!;
+                if (Math.abs(left - scrollLeft) >= 1 || Math.abs(top - scrollTop) >= 1) {
+                    this.setScrollLeft(scrollLeft);
+                    this.setScrollTop(scrollTop);
+                    this.setViewport();
+                }
             });
         this.resizeElement();
     }
@@ -344,253 +309,218 @@ export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, On
     resizeElement() {
         this.resizeObserver = new ResizeObserver(entries => {
             for (let entry of entries) {
-                const { width, height } = entry.contentRect;
-                const hideScrollbar = this.board.options.hideScrollbar;
-                const scrollBarWidth = hideScrollbar ? SCROLL_BAR_WIDTH : 0;
-                this.width = width + scrollBarWidth;
-                this.height = height + scrollBarWidth;
-                this.changeSize();
+                this.initViewportContainer();
+                this.calcViewBox(this.board.viewport.zoom);
+                this.updateViewBoxStyles();
+                this.updateViewportScrolling();
+                this.setViewport();
             }
         });
         this.resizeObserver.observe(this.elementRef.nativeElement);
     }
 
-    initContainerSize() {
-        const clientBox = getViewportClientBox(this.board);
-        this.width = clientBox.width;
-        this.height = clientBox.height;
-        this.resizeViewport();
+    private updateViewportState(state: Partial<PlaitBoardViewport>) {
+        this.viewportState = {
+            ...this.viewportState,
+            ...state
+        };
     }
 
-    resizeViewport() {
-        const container = this.elementRef.nativeElement?.parentElement;
-        const containerRect = container?.getBoundingClientRect();
-        const hideScrollbar = this.board.options.hideScrollbar;
-        const scrollBarWidth = hideScrollbar ? SCROLL_BAR_WIDTH : 0;
-        const width = `${containerRect.width + scrollBarWidth}px`;
-        const height = `${containerRect.height + scrollBarWidth}px`;
-
-        this.renderer2.setStyle(this.contentContainer.nativeElement, 'width', width);
-        this.renderer2.setStyle(this.contentContainer.nativeElement, 'height', height);
-        this.renderer2.setStyle(this.contentContainer.nativeElement, 'maxWidth', width);
-        this.renderer2.setStyle(this.contentContainer.nativeElement, 'maxHeight', height);
+    initViewportContainer() {
+        const { width, height } = getBoardClientBox(this.board);
+        this.renderer2.setStyle(this.viewportContainer.nativeElement, 'width', `${width}px`);
+        this.renderer2.setStyle(this.viewportContainer.nativeElement, 'height', `${height}px`);
     }
 
-    setMatrix() {
-        const viewBox = this.viewBox;
-        const zoom = this.zoom;
-        this.matrix = [zoom, 0, 0, 0, zoom, 0, -this.scrollLeft - zoom * viewBox[0], -this.scrollTop - zoom * viewBox[1], 1];
-    }
+    initViewport(viewport = this.board.viewport) {
+        const originationCoord = viewport?.originationCoord;
+        const zoom = viewport?.zoom ?? this.viewportState.zoom;
 
-    updateViewport() {
-        const clientBox = getViewportClientBox(this.board);
-        this.setViewport(this.board.viewport.zoom, [clientBox.x, clientBox.y]);
-    }
-
-    setViewport(zoom?: number, focusPoint?: number[]) {
-        zoom = zoom ?? this.zoom;
-        zoom = calculateZoom(zoom);
-        if (zoom !== this.zoom) {
-            this.zoom = zoom;
+        if (originationCoord) {
+            const matrix = this.getMatrix();
+            const [pointX, pointY] = invertViewportCoordinates([0, 0], matrix);
+            const scrollLeft = this.viewportState.scrollLeft!;
+            const scrollTop = this.viewportState.scrollTop!;
+            const left = scrollLeft + (originationCoord[0] - pointX) * zoom;
+            const top = scrollTop + (originationCoord[1] - pointY) * zoom;
+            this.setScroll(left, top);
+        } else {
+            this.adaptHandle();
         }
+    }
 
-        focusPoint = focusPoint ?? this.focusPoint;
+    calcViewBox(zoom = this.viewportState.zoom) {
+        zoom = clampZoomLevel(zoom);
 
+        const scrollBarWidth = this.plaitOptions?.hideScrollbar ? SCROLL_BAR_WIDTH : 0;
+        const viewportContainerBox = getViewportContainerBox(this.board);
+        const groupBBox = getRootGroupBBox(this.board, zoom);
+        const horizontalPadding = viewportContainerBox.width / 2;
+        const verticalPadding = viewportContainerBox.height / 2;
+        const viewportWidth = (groupBBox.right - groupBBox.left) * zoom + 2 * horizontalPadding + scrollBarWidth;
+        const viewportHeight = (groupBBox.bottom - groupBBox.top) * zoom + 2 * verticalPadding + scrollBarWidth;
+
+        const viewBox = [
+            groupBBox.left - horizontalPadding / zoom,
+            groupBBox.top - verticalPadding / zoom,
+            viewportWidth / zoom,
+            viewportHeight / zoom
+        ];
+        const matrix = this.getMatrix();
         let scrollLeft;
         let scrollTop;
-        const clientBox = getViewportClientBox(this.board);
-        const matrix = this.matrix;
-        const box = calculateBBox(this.board, zoom);
-        const padding = [this.height / 2, this.width / 2];
-        const rootGroupWidth = box.right - box.left;
-        const rootGroupHeight = box.bottom - box.top;
-        const canvasWidth = rootGroupWidth * zoom + 2 * padding[1];
-        const canvasHeight = rootGroupHeight * zoom + 2 * padding[0];
-        const viewBox = [box.left - padding[1] / zoom, box.top - padding[0] / zoom, canvasWidth / zoom, canvasHeight / zoom];
 
-        if (matrix) {
-            const canvasPoint = [focusPoint[0] - clientBox.x, focusPoint[1] - clientBox.y, 1];
-            const invertMatrix = invert([], matrix);
-            const matrix1 = transformMat3([], [canvasPoint[0], canvasPoint[1], 1], invertMatrix as []);
-            const matrix2 = [zoom, 0, 0, 0, zoom, 0, -zoom * viewBox[0], -zoom * viewBox[1], 1];
-            const newMatrix = transformMat3([], matrix1, matrix2);
+        if (matrix.length > 0) {
+            // focusPoint
+            const focusX = viewportContainerBox.x + viewportContainerBox.width / 2;
+            const focusY = viewportContainerBox.y + viewportContainerBox.height / 2;
+            const viewportContainerPoint = [focusX - viewportContainerBox.x, focusY - viewportContainerBox.y, 1];
+            const point = invertViewportCoordinates([viewportContainerPoint[0], viewportContainerPoint[1]], matrix);
+            const newMatrix = [zoom, 0, 0, 0, zoom, 0, -zoom * viewBox[0], -zoom * viewBox[1], 1];
+            const newPoint = transformMat3([], point, newMatrix);
 
-            scrollLeft = newMatrix[0] - canvasPoint[0];
-            scrollTop = newMatrix[1] - canvasPoint[1];
+            scrollLeft = newPoint[0] - viewportContainerPoint[0];
+            scrollTop = newPoint[1] - viewportContainerPoint[1];
         } else {
-            scrollLeft = (canvasWidth - clientBox.width) / 2;
-            scrollTop = padding[0] / 2 - box.top;
+            scrollLeft = horizontalPadding;
+            scrollTop = verticalPadding;
         }
-
-        this.canvasWidth = canvasWidth;
-        this.canvasHeight = canvasHeight;
-        this.zoom = zoom;
-        this.viewBox = viewBox;
+        this.updateViewportState({
+            viewportWidth,
+            viewportHeight,
+            zoom,
+            viewBox
+        });
         this.setScrollLeft(scrollLeft);
         this.setScrollTop(scrollTop);
-        this.change();
+    }
+
+    getMatrix(): number[] {
+        const { viewBox, zoom, scrollLeft, scrollTop } = this.viewportState;
+        if (scrollLeft! >= 0 && scrollTop! >= 0) {
+            return [zoom, 0, 0, 0, zoom, 0, -scrollLeft! - zoom * viewBox![0], -scrollTop! - zoom * viewBox![1], 1];
+        }
+        return [];
     }
 
     setScrollLeft(left: number) {
-        const hideScrollbar = this.board.options.hideScrollbar;
-        const scrollBarWidth = hideScrollbar ? SCROLL_BAR_WIDTH : 0;
-        const width = this.canvasWidth - this.width + scrollBarWidth;
-        this.scrollLeft = left < 0 ? 0 : left > width ? width : left;
+        const viewportContainerBox = getViewportContainerBox(this.board);
+        const scrollBarWidth = this.plaitOptions?.hideScrollbar ? SCROLL_BAR_WIDTH : 0;
+        const { viewportWidth } = this.viewportState;
+        const width = viewportWidth! - viewportContainerBox.width + scrollBarWidth;
+        this.viewportState.scrollLeft = left < 0 ? 0 : left > width ? width : left;
     }
 
     setScrollTop(top: number) {
-        const hideScrollbar = this.board.options.hideScrollbar;
-        const scrollBarWidth = hideScrollbar ? SCROLL_BAR_WIDTH : 0;
-        const height = this.canvasHeight - this.height + scrollBarWidth;
-        this.scrollTop = top < 0 ? 0 : top > height ? height : top;
+        const viewportContainerBox = getViewportContainerBox(this.board);
+        const scrollBarWidth = this.plaitOptions?.hideScrollbar ? SCROLL_BAR_WIDTH : 0;
+        const { viewportHeight } = this.viewportState;
+        const height = viewportHeight! - viewportContainerBox.height + scrollBarWidth;
+        this.viewportState.scrollTop = top < 0 ? 0 : top > height ? height : top;
     }
 
     setScroll(left: number, top: number) {
         this.setScrollLeft(left);
         this.setScrollTop(top);
-        this.change();
+        this.updateViewBoxStyles();
+        this.updateViewportScrolling();
+        this.setViewport();
     }
 
-    moveTo(point: number[], zoom: number) {
-        zoom = zoom ?? this.zoom;
-        const r = invertViewport([0, 0], this.matrix);
-        this.setScroll(this.scrollLeft + (point[0] - r[0]) * zoom, this.scrollTop + (point[1] - r[1]) * zoom);
-    }
+    private updateViewBoxStyles() {
+        const { host, viewportState } = this;
+        const { viewportWidth, viewportHeight, viewBox } = viewportState;
 
-    scrollIntoView(node: { x: number; y: number; width: number; height: number }) {
-        const canvasRect = this.host.getBoundingClientRect();
+        this.renderer2.setStyle(host, 'display', 'block');
+        this.renderer2.setStyle(host, 'width', `${viewportWidth}px`);
+        this.renderer2.setStyle(host, 'height', `${viewportHeight}px`);
 
-        if (!(canvasRect.width <= this.width && canvasRect.height <= this.height)) {
-            const point = convertViewport([node.x, node.y], this.matrix);
-            const fullPoint = convertViewport([node.x + node.width, node.y + node.height], this.matrix);
-            const width = this.width;
-            const height = this.height;
-            let left = this.scrollLeft;
-            let top = this.scrollTop;
-
-            if (fullPoint[0] > width - SCROLL_BAR_WIDTH) {
-                left += fullPoint[0] - width + SCROLL_BAR_WIDTH;
-            }
-            if (fullPoint[1] > height - SCROLL_BAR_WIDTH) {
-                top += fullPoint[1] - height + SCROLL_BAR_WIDTH;
-            }
-            (left === this.scrollLeft && top === this.scrollTop) || this.setScroll(left, top);
+        if (viewBox && viewBox[2] > 0 && viewBox[3] > 0) {
+            this.renderer2.setAttribute(host, 'viewBox', viewBox.join(' '));
         }
     }
 
-    change() {
-        this.resizeViewport();
-        this.renderer2.setStyle(this.host, 'display', 'block');
-        this.renderer2.setStyle(this.host, 'width', `${this.canvasWidth}px`);
-        this.renderer2.setStyle(this.host, 'height', `${this.canvasHeight}px`);
-        if (this.viewBox && this.viewBox[2] > 0 && this.viewBox[3] > 0) {
-            this.renderer2.setAttribute(this.host, 'viewBox', this.viewBox.join(','));
-        }
-
-        this.contentContainer.nativeElement.scrollLeft = this.scrollLeft;
-        this.contentContainer.nativeElement.scrollTop = this.scrollTop;
-
-        this.setMatrix();
-        this.setViewportSetting();
-        this.viewportCanvasBox = getViewportCanvasBox(this.board, this.matrix);
+    private updateViewportScrolling() {
+        const { viewportContainer, viewportState } = this;
+        const { scrollLeft, scrollTop } = viewportState;
+        viewportContainer.nativeElement.scrollLeft = scrollLeft!;
+        viewportContainer.nativeElement.scrollTop = scrollTop!;
     }
 
-    restoreCanvasPoint(point: number[], zoom?: number) {
-        zoom = zoom ?? this.zoom;
-        this.setViewport(zoom);
-        this.moveTo(point, zoom);
-    }
-
-    initCanvas(viewport: Viewport) {
-        const canvasPoint = viewport?.canvasPoint;
-
-        if (canvasPoint) {
-            this.restoreCanvasPoint(canvasPoint, viewport.zoom);
-        } else {
-            this.setViewport(viewport.zoom);
-            this.setScroll(this.scrollLeft, this.scrollTop);
-        }
-    }
-
-    focusGraphicCenter() {
-        const rootGroup = this.board.host.firstChild;
-        const rootGroupBox = (rootGroup as SVGGraphicsElement).getBBox();
-        this.focus([rootGroupBox.x + rootGroupBox.width / 2, rootGroupBox.y + rootGroupBox.height / 2]);
-        this.change();
-        this.setViewport(1);
-    }
-
-    focus(point: number[]) {
-        const clientBox = getViewportClientBox(this.board);
-        const matrix = transformMat3([], [point[0], point[1], 1], this.matrix);
-        const newPoint = [clientBox.width / 2, clientBox.height / 2];
-        const scrollLeft = newPoint[0] - matrix[0];
-        const scrollTop = newPoint[1] - matrix[1];
-
-        this.setScrollLeft(this.scrollLeft - scrollLeft);
-        this.setScrollTop(this.scrollTop - scrollTop);
-        this.setMatrix();
-    }
-
-    resetFocusPoint() {
-        const clientBox = getViewportClientBox(this.board);
-        this.focusPoint = [clientBox.x + clientBox.width / 2, clientBox.y + clientBox.height / 2];
-    }
-
-    setViewportSetting() {
+    setViewport() {
         const viewport = this.board?.viewport;
-        Transforms.setViewport(this.board, {
-            ...viewport,
-            zoom: this.zoom,
-            canvasPoint: invertViewport([0, 0], this.matrix)
-        });
-    }
+        const oldOriginationCoord = viewport?.originationCoord ?? [];
+        const matrix = this.getMatrix();
+        const originationCoord = invertViewportCoordinates([0, 0], matrix);
 
-    changeSize() {
-        this.updateViewport();
+        if (!originationCoord.every((item, index) => item === oldOriginationCoord[index])) {
+            Transforms.setViewport(this.board, {
+                ...viewport,
+                zoom: this.viewportState.zoom,
+                originationCoord
+            });
+        }
     }
 
     trackBy = (index: number, element: PlaitElement) => {
         return index;
     };
 
-    // 拖拽模式
-    changeMoveMode(cursorStatus: BaseCursorStatus) {
-        updateCursorStatus(this.board, cursorStatus);
-        this.cdr.markForCheck();
-    }
-
-    // 适应画布
     adaptHandle() {
-        const clientBox = getViewportClientBox(this.board);
-        const rootGroup = this.host.firstChild;
-        const rootGroupBox = (rootGroup as SVGGraphicsElement).getBBox();
-        const viewportWidth = clientBox.width - 2 * this.autoFitPadding;
-        const viewportHeight = clientBox.height - 2 * this.autoFitPadding;
+        const containerBox = getViewportContainerBox(this.board);
+        const rootGroup = this.host.firstChild as SVGGraphicsElement;
+        const matrix = this.getMatrix();
 
-        let zoom = this.zoom;
+        const rootGroupBox = rootGroup.getBBox();
+        const centerPoint = [containerBox.width / 2, containerBox.height / 2];
+        const rootGroupCenterX = rootGroupBox.x + rootGroupBox.width / 2;
+        const rootGroupCenterY = rootGroupBox.y + rootGroupBox.height / 2;
+        const transformedPoint = transformMat3([], [rootGroupCenterX, rootGroupCenterY, 1], matrix);
+
+        const offsetLeft = centerPoint[0] - transformedPoint[0];
+        const offsetTop = centerPoint[1] - transformedPoint[1];
+
+        const { autoFitPadding, zoom, scrollLeft, scrollTop } = this.viewportState;
+        const viewportWidth = containerBox.width - 2 * autoFitPadding;
+        const viewportHeight = containerBox.height - 2 * autoFitPadding;
+        let newZoom = zoom;
         if (viewportWidth < rootGroupBox.width || viewportHeight < rootGroupBox.height) {
-            zoom = Math.min(viewportWidth / rootGroupBox.width, viewportHeight / rootGroupBox.height);
+            newZoom = Math.min(viewportWidth / rootGroupBox.width, viewportHeight / rootGroupBox.height);
         } else {
-            zoom = 1;
+            newZoom = 1;
         }
-        this.focus([rootGroupBox.x + rootGroupBox.width / 2, rootGroupBox.y + rootGroupBox.height / 2]);
-        this.resetFocusPoint();
-        this.setViewport(zoom);
+
+        this.setScrollLeft(scrollLeft! - offsetLeft);
+        this.setScrollTop(scrollTop! - offsetTop);
+        this.calcViewBox(newZoom);
+        this.updateViewBoxStyles();
+        this.updateViewportScrolling();
+        this.setViewport();
     }
 
-    // 放大
     zoomInHandle() {
-        this.setViewport(this.zoom + 0.1);
+        this.calcViewBox(this.viewportState.zoom + 0.1);
+        this.updateViewBoxStyles();
+        this.updateViewportScrolling();
+        this.setViewport();
     }
 
-    // 缩小
     zoomOutHandle() {
-        this.setViewport(this.zoom - 0.1);
+        this.calcViewBox(this.viewportState.zoom - 0.1);
+        this.updateViewBoxStyles();
+        this.updateViewportScrolling();
+        this.setViewport();
     }
 
     resetZoomHandel() {
-        this.setViewport(1);
+        this.calcViewBox(1);
+        this.updateViewBoxStyles();
+        this.updateViewportScrolling();
+        this.setViewport();
+    }
+
+    movingChange(isMoving: boolean) {
+        this.isMoving = isMoving;
     }
 
     ngOnDestroy(): void {
@@ -602,7 +532,48 @@ export class PlaitBoardComponent implements OnInit, OnChanges, AfterViewInit, On
         }
     }
 
-    movingChange(isMoving: boolean) {
-        this.isMoving = isMoving;
+    markForCheck() {
+        this.cdr.markForCheck();
+    }
+
+    scrollToRectangle(client: RectangleClient) {
+        this.calcViewBox();
+        this.updateViewBoxStyles();
+        this.updateViewportScrolling();
+        this.setViewport();
+
+        const svgRect = this.host.getBoundingClientRect();
+        const viewportContainerBox = getViewportContainerBox(this.board);
+
+        if (svgRect.width > viewportContainerBox.width || svgRect.height > viewportContainerBox.height) {
+            const scrollBarWidth = this.plaitOptions?.hideScrollbar ? SCROLL_BAR_WIDTH : 0;
+            const matrix = this.getMatrix();
+            const [nodePointX, nodePointY] = convertToViewportCoordinates([client.x, client.y], matrix);
+            const [fullNodePointX, fullNodePointY] = convertToViewportCoordinates(
+                [client.x + client.width, client.y + client.height],
+                matrix
+            );
+
+            let newLeft = this.viewportState.scrollLeft!;
+            let newTop = this.viewportState.scrollTop!;
+
+            if (nodePointX < 0) {
+                newLeft -= Math.abs(nodePointX);
+            }
+            if (nodePointX > 0 && fullNodePointX > viewportContainerBox.width) {
+                newLeft += fullNodePointX - viewportContainerBox.width + scrollBarWidth;
+            }
+
+            if (nodePointY < 0) {
+                newTop -= Math.abs(nodePointY);
+            }
+            if (nodePointY > 0 && fullNodePointY > viewportContainerBox.height) {
+                newTop += fullNodePointY - viewportContainerBox.height + scrollBarWidth;
+            }
+
+            if (newLeft !== this.viewportState.scrollLeft! || newTop !== this.viewportState.scrollTop!) {
+                this.setScroll(newLeft, newTop);
+            }
+        }
     }
 }