+++ packages/plait/src/plugins/with-selection.ts
@@ -1,53 +1,57 @@
 import { PlaitBoard } from '../interfaces/board';
 import { Point } from '../interfaces/point';
 import { Transforms } from '../transforms';
-import { isNoSelectionElement } from '../utils/board';
+import { transformPoint } from '../utils/board';
 import { toPoint } from '../utils/dom';
-import { toRectangleClient } from '../utils/graph';
+import { RectangleClient } from '../interfaces/rectangle-client';
+import { cacheSelectedElements, calcElementIntersectionSelection } from '../utils/selected-element';
 
 export function withSelection<T extends PlaitBoard>(board: T) {
-    const { mousedown, mousemove, globalMouseup } = board;
+    const { mousedown, mousemove, mouseup, onChange } = board;
 
     let start: Point | null = null;
     let end: Point | null = null;
 
     board.mousedown = (event: MouseEvent) => {
-        if (isNoSelectionElement(event)) {
-            mousedown(event);
-            return;
-        }
-        if (event.target instanceof Node && board.host.contains(event.target)) {
-            start = toPoint(event.x, event.y, board.host);
-        }
+        start = transformPoint(board, toPoint(event.x, event.y, board.host));
         mousedown(event);
     };
 
     board.mousemove = (event: MouseEvent) => {
-        const movedTarget = toPoint(event.x, event.y, board.host);
+        const movedTarget = transformPoint(board, toPoint(event.x, event.y, board.host));
         if (start) {
-            const rectangleClient = toRectangleClient([start, movedTarget]);
-            if (start && Math.hypot(rectangleClient.width, rectangleClient.height) > 5) {
+            const rectangleClient = RectangleClient.toRectangleClient([start, movedTarget]);
+            if (Math.hypot(rectangleClient.width, rectangleClient.height) > 5) {
                 end = movedTarget;
             }
         }
         mousemove(event);
     };
 
-    board.globalMouseup = (event: MouseEvent) => {
-        if (isNoSelectionElement(event)) {
-            return globalMouseup(event);
-        }
-        if (start) {
+    board.mouseup = (event: MouseEvent) => {
+        if (start && end) {
+            Transforms.setSelection(board, { anchor: start, focus: end });
+        } else if (start) {
             Transforms.setSelection(board, { anchor: start, focus: start });
-        } else {
-            if (board.selection !== null) {
-                Transforms.setSelection(board, null);
-            }
         }
+
         start = null;
         end = null;
 
-        globalMouseup(event);
+        mouseup(event);
+    };
+
+    board.onChange = () => {
+        // calc selected elements entry
+        try {
+            if (board.operations.find(value => value.type === 'set_selection')) {
+                const elementIds = calcElementIntersectionSelection(board);
+                cacheSelectedElements(board, elementIds);
+            }
+        } catch (error) {
+            console.error(error);
+        }
+        onChange();
     };
 
     return board;